genai/travelbuddy_ai/pyproject.toml
[project]
name = "travel_buddy_ai"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "langchain-openai (>=0.3.24,<0.4.0)",
    "qdrant-client (>=1.14.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
travel-buddy-api = "travel_buddy_ai.main:create_app"